rag-qwen3/requirements.txt
sentence-transformers
faiss-cpu
llama-cpp-python
